+++ src/application.rs
@@ -7,18 +7,15 @@ use winit::event::{DeviceEvent, DeviceId, WindowEvent};
 use winit::event_loop::{ActiveEventLoop, EventLoopClosed, EventLoopProxy};
 use winit::window::{WindowAttributes, WindowId};
 
-use control_overlay::ControlState;
-
 use crate::build::tenscript::{FabricPlan, TenscriptError};
 use crate::build::tenscript::brick_library::BrickLibrary;
 use crate::build::tenscript::fabric_library::FabricLibrary;
-use crate::control_overlay;
-use crate::control_overlay::action::Action;
 use crate::crucible::{Crucible, CrucibleAction};
+use crate::messages::{ControlState, LabEvent};
 use crate::scene::Scene;
 use crate::wgpu::Wgpu;
 
-pub struct Application {
+ pub struct Application {
     window_attributes: WindowAttributes,
     scene: Option<Scene>,
     crucible: Crucible,
@@ -29,27 +26,27 @@ pub struct Application {
     pub brick_library: BrickLibrary,
     pub control_state: ReadSignal<ControlState>,
     pub set_control_state: WriteSignal<ControlState>,
-    pub event_loop_proxy: Arc<EventLoopProxy<Action>>,
+    pub event_loop_proxy: Arc<EventLoopProxy<LabEvent>>,
 }
 
-impl ApplicationHandler<Action> for Application {
+impl ApplicationHandler<LabEvent> for Application {
     fn resumed(&mut self, event_loop: &ActiveEventLoop) {
         let window = Arc::new(event_loop.create_window(self.window_attributes.clone()).unwrap());
         let event_loop_proxy = self.event_loop_proxy.clone();
         Wgpu::create_and_send(window, event_loop_proxy);
     }
 
-    fn user_event(&mut self, _event_loop: &ActiveEventLoop, event: Action) {
+    fn user_event(&mut self, _event_loop: &ActiveEventLoop, event: LabEvent) {
         match event {
-            Action::ContextCreated(wgpu) => {
+            LabEvent::ContextCreated(wgpu) => {
                 let scene = Scene::new(wgpu, (self.control_state, self.set_control_state));
                 self.scene = Some(scene)
             }
-            Action::LoadFabric(fabric_plan_name) => {
+            LabEvent::LoadFabric(fabric_plan_name) => {
                 self.fabric_plan_name = fabric_plan_name;
                 self.reload_fabric();
             }
-            Action::Crucible(crucible_action) => {
+            LabEvent::Crucible(crucible_action) => {
                 if let CrucibleAction::BuildFabric(fabric_plan) = &crucible_action {
                     self.fabric_plan_name = fabric_plan.name.clone();
                     if let Some(scene) = &mut self.scene {
@@ -59,21 +56,21 @@ impl ApplicationHandler<Action> for Application {
                 self.crucible.action(crucible_action);
             }
             #[cfg(target_arch = "wasm32")]
-            Action::UpdatedLibrary(_) => unreachable!(),
+            LabEvent::UpdatedLibrary(_) => unreachable!(),
             #[cfg(not(target_arch = "wasm32"))]
-            Action::UpdatedLibrary(time) => {
+            LabEvent::UpdatedLibrary(time) => {
                 let _fabric_library = self.fabric_library.clone();
                 self.fabric_library_modified = time;
                 if !self.fabric_plan_name.is_empty() {
                     self.reload_fabric();
                 }
             }
-            Action::Scene(pick) => {
+            LabEvent::Scene(pick) => {
                 if let Some(scene) = &mut self.scene {
                     scene.do_pick(pick);
                 }
             }
-            Action::CalibrateStrain => {
+            LabEvent::CalibrateStrain => {
                 // let strain_limits =
                 //     self.crucible.fabric().strain_limits(":bow-tie".to_string());
                 // self.user_interface.set_strain_limits(strain_limits);
@@ -120,7 +117,7 @@ impl Application {
     pub fn new(
         window_attributes: WindowAttributes,
         (control_state, set_control_state): (ReadSignal<ControlState>, WriteSignal<ControlState>),
-        event_loop_proxy: Arc<EventLoopProxy<Action>>,
+        event_loop_proxy: Arc<EventLoopProxy<LabEvent>>,
     ) -> Result<Application, TenscriptError> {
         let brick_library = BrickLibrary::from_source()?;
         let fabric_library = FabricLibrary::from_source()?;
@@ -174,7 +171,7 @@ impl Application {
         }
     }
 
-    pub fn build_fabric(&mut self, fabric_name: &String) -> Result<(), EventLoopClosed<Action>> {
+    pub fn build_fabric(&mut self, fabric_name: &String) -> Result<(), EventLoopClosed<LabEvent>> {
         let fabric_plan = self
             .fabric_library
             .fabric_plans
@@ -182,7 +179,7 @@ impl Application {
             .find(|FabricPlan { name, .. }| name == fabric_name)
             .expect(fabric_name);
         self.event_loop_proxy
-            .send_event(Action::Crucible(CrucibleAction::BuildFabric(fabric_plan.clone())))
+            .send_event(LabEvent::Crucible(CrucibleAction::BuildFabric(fabric_plan.clone())))
     }
 
     pub fn capture_prototype(&mut self, brick_index: usize) {
@@ -196,9 +193,9 @@ impl Application {
         self.crucible.action(CrucibleAction::BakeBrick(prototype));
     }
 
-    pub fn refresh_library(&mut self, time: SystemTime) -> Result<Action, TenscriptError> {
+    pub fn refresh_library(&mut self, time: SystemTime) -> Result<LabEvent, TenscriptError> {
         self.fabric_library = FabricLibrary::from_source()?;
-        Ok(Action::UpdatedLibrary(time))
+        Ok(LabEvent::UpdatedLibrary(time))
     }
 
     pub fn load_preset(&self, plan_name: String) -> Result<FabricPlan, TenscriptError> {
