--- src/control_state.rs
@@ -1,19 +1,0 @@
-use crate::fabric::interval::Role;
-
-#[derive(Clone, Debug, Copy)]
-pub struct IntervalDetails {
-    pub alpha_index: usize,
-    pub omega_index: usize,
-    pub length: f32,
-    pub role: Role,
-}
-
-#[derive(Clone, Debug, Default)]
-pub enum ControlState {
-    #[default]
-    Choosing,
-    Viewing,
-    ShowingJoint(usize),
-    ShowingInterval(IntervalDetails),
-    SettingLength(IntervalDetails),
-}
